+++ CHANGELOG.md
@@ -87,6 +87,7 @@
 * pool: event verification cache ([Yuki Kishimoto])
 * pool: add `AdmitPolicy` trait ([Yuki Kishimoto])
 * ffi: add Mac Catalyst support in Swift package ([Yuki Kishimoto])
+* js: add `KindStandard` enum ([Yuki Kishimoto])
 
 ### Fixed